refactor(#1): 불필요한 Component directory 제거
