+++ src/components/common/CheckInput.tsx
